Remove nonfunctional [code_]libraries_gui support (#1639)
