+++ lang/python/python.talon
@@ -11,7 +11,6 @@ tag(): user.code_functions
 tag(): user.code_functions_common
 tag(): user.code_keywords
 tag(): user.code_libraries
-tag(): user.code_libraries_gui
 tag(): user.code_operators_array
 tag(): user.code_operators_assignment
 tag(): user.code_operators_bitwise
