+++ lang/rust/rust.py
@@ -178,7 +178,7 @@ all_traits = {
 }
 
 
-# tag: libraries_gui
+# tag: libraries
 ctx.lists["user.code_libraries"] = {
     "eye oh": "std::io",
     "file system": "std::fs",
@@ -345,8 +345,6 @@ class UserActions:
 
     def code_import():
         actions.auto_insert("use ")
-
-    # tag: libraries_gui
 
     def code_insert_library(text: str, selection: str):
         actions.user.paste(f"use {text}")
