--- lang/tags/libraries_gui.talon
@@ -1,8 +1,0 @@
-tag: user.code_libraries_gui
--
-# NOTE: This file does not define any commands, as the commands vary from
-#       language to language, e.g., Python uses 'import', C uses 'include',
-#       R uses 'library', etcetera.
-
-# TODO: If this ever becomes possible, we should abstract over these commands
-#       using a variable which can be set to the context-specific word.
